pyproject.toml
[tool.poetry]
name = "gif-secret"
version = "0.1.0"
description = ""
authors = ["Bentsi Magidovich <magidovich@gmail.com>"]
readme = "README.md"
packages = [{include = "gif_secret"}]

[tool.poetry.dependencies]
python = "^3.11"
pillow = "^9.5.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
